pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "snowbox"
version = "0.2.0"
description = "FastMCP server: Snowflake-connected Python in auditable Marimo notebooks"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Fahad Baig", email = "mfbaig35r@gmail.com"},
]
dependencies = [
    "marimo-sandbox>=1.0.0",
    "fastmcp>=2.0.0",
    "snowflake-connector-python[pandas]>=3.0.0",
    "pyarrow>=14.0",
    "pandas>=2.0",
    "pydantic>=2.0",
]

[project.urls]
Homepage = "https://github.com/mfbaig35r/snowbox"
Repository = "https://github.com/mfbaig35r/snowbox"

[project.optional-dependencies]
dev = ["pytest>=7", "pytest-timeout", "ruff", "mypy>=1.0"]

[project.scripts]
snowbox = "snowbox.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/snowbox"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
strict = false
disallow_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["snowflake.*", "fastmcp.*", "marimo.*", "marimo_sandbox.*", "pyarrow.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["slow: integration tests that require real Snowflake credentials"]
